Player disconnect and rejoin seem to work now
Also removed a state that was not needed. SHOW_STANDINGS is now the final state, and we wrap at SHOW_CORRECT.
Moved MIN_INF constant into the const files...
Removed a lot of console logging
Bruker mer norsk i tekst :) TODO: make translation file or something.
Redesigned client-networking: now a giant "state-update" message, but only parts of it is be used for some of the updates. Send functions moved to player "class". This was because I needed to send more state to player when reconnect happens.