Split Host and Player by simple copy paste code
And then only some small changes to each so that the "enter credentials" page for host/player are the front page for each of the two apps.
Deployed and seem to work :)
TODO: see if I can move shared code to some location both apps can reach, so I dont have to maintain duplicates (e.g. of css and some helper functions)